src/impulsoprevine/requirements.txt
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.7
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
bleach==3.3.0
blinker==1.4
bs4==0.0.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==36.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
entrypoints==0.3
filelock==3.0.12
future==0.18.2
geocoder==1.38.1
gitdb==4.0.7
GitPython==3.1.14
greenlet==1.1.1
h11==0.12.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.36.0
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numba==0.53.1
numpy==1.20.1
outcome==1.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.6
psycopg2==2.9.1
pyarrow==4.0.0
pycodestyle==2.7.0
pycparser==2.20
pydeck==0.6.1
Pygments==2.8.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
PySys==1.6.1
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2021.1
# pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
ratelim==0.1.6
requests==2.25.1
selenium==4.1.0
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
streamlit==0.79.0
terminado==0.9.4
testpath==0.4.4
thrift==0.13.0
timestring==1.6.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
trio==0.19.0
trio-websocket==0.9.2
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.4
validators==0.18.2
virtualenv==20.4.3
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wsproto==1.0.0
xlrd==2.0.1
